Update index.html
It is improved, instead of it being a check in the convert option, there should simply be the numeric value input field. When you enter one, the other is blocked only to do the conversion and if you want to go to the other text input, you first have to delete what you had written and then write the corresponding option, in order to save computational effort.